third_party/blink/web_tests/external/wpt/css/motion/parsing/offset-path-parsing-invalid-expected.txt

This is a testharness.js-based test.
PASS e.style['offset-path'] = "path(\"M 20 30 A 60 70 80\")" should not set the property value
FAIL e.style['offset-path'] = "path(\"\")" should not set the property value assert_equals: expected "" but got "none"
FAIL e.style['offset-path'] = "path(\" \")" should not set the property value assert_equals: expected "" but got "none"
PASS e.style['offset-path'] = "ray(0 sides)" should not set the property value
PASS e.style['offset-path'] = "ray(closest-side)" should not set the property value
PASS e.style['offset-path'] = "ray(closest-side 0deg closest-side)" should not set the property value
PASS e.style['offset-path'] = "ray(0deg closest-side 0deg)" should not set the property value
PASS e.style['offset-path'] = "ray(contain 0deg closest-side contain)" should not set the property value
Harness: the test ran to completion.
